+++ src/widgets/mod.rs
@@ -17,3 +17,9 @@
 
 /// This is a `TextWidget`, which draws text in a clipped area.
 pub mod text_widget;
+
+/// This is an `ImageWidget`, which draws an Image in a clipped area.
+pub mod image_widget;
+
+/// This is a `ProgressWidget`, which draws a progress bar.
+pub mod progress_widget;