+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `ImageWidget`.
 - Added `ImagePosition` to specify the position of the image inside the `Widget`.
 - Added `image` example to show all coordinates for the image and scaling examples.
+- Removed old text and image widgets.
 
 ## 0.4.7