+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Moved `Widget` macros to top-level `lib.rs` so that it can be shared between modules.
 - Modified `TextWidget` to use both macros.
-- Added `ImageWidget`.
+- Added `ImageWidget` (#203)
 - Added `ImagePosition` to specify the position of the image inside the `Widget`.
 - Added `image` example to show all coordinates for the image and scaling examples.
 - Removed old text and image widgets.