+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-# Pushrod Releases
+# Pushrod Change Log
+
+## 0.4.8
+
+- Moved `Widget` macros to top-level `lib.rs` so that it can be shared between modules.
+- Modified `TextWidget` to use both macros.
+- Added `ImageWidget` (#203)
+- Added `ImagePosition` to specify the position of the image inside the `Widget`.
+- Added `image` example to show all coordinates for the image and scaling examples.
+- Removed old text and image widgets.
+- Ported `ProgressWidget` (#204)
+- Added `progress` demo to show progress widgets.
 
 ## 0.4.7